% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.spBFA}
\alias{is.spBFA}
\title{is.spBFA}
\usage{
is.spBFA(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.spBFA} is a general test of an object being interpretable as a
\code{\link{spBFA}} object.
}
\details{
The \code{\link{spBFA}} class is defined as the regression object that
 results from the \code{\link{spBFA}} regression function.
}
\examples{
###Load pre-computed results
data(reg.bfa_sp)

###Test function
is.spBFA(reg.bfa_sp)

}
