% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.bfa_sp-data.R
\docType{data}
\name{reg.bfa_sp}
\alias{reg.bfa_sp}
\title{Pre-computed regression results from \code{bfa_sp}}
\format{The data object \code{reg.bfa_sp} is a \code{spBFA} data object that is the result of implementing the MCMC code in the vignette.
        It is presented here because the run-time would be unecessarily long when compiling the R package.}
\usage{
data(reg.bfa_sp)
}
\description{
The data object \code{reg.bfa_sp} is a pre-computed \code{spBFA} data object for use in the package vignette and function examples.
}
\keyword{datasets}
