% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spBFA-package.R
\docType{package}
\name{spBFA}
\alias{spBFA-package}
\alias{spBFA}
\title{spBFA}
\description{
\code{spBFA} is a package for Bayesian spatial factor analysis. A corresponding manuscript is forthcoming.
}
\author{
Samuel I. Berchuck \email{sib2@duke.edu}
}
