% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRED_predict.R
\name{predict.spBFA}
\alias{predict.spBFA}
\title{predict.spBFA}
\usage{
\method{predict}{spBFA}(object, NewTimes, NewX = NULL,
  NewTrials = NULL, type = "temporal", ...)
}
\arguments{
\item{object}{A \code{\link{spBFA}} model object for which predictions
are desired from.}

\item{NewTimes}{A numeric vector including desired time(s) points for prediction.}

\item{NewX}{A matrix including covariates at times \code{NewTimes} for prediction. 
\code{NewX} must have dimension \code{(M x O x NNewVistis) x P}. Where \code{NNewVisits} is the number of temporal 
locations being predicted. The default sets \code{NewX} to \code{NULL}, which assumes that the covariates for all predictions 
are the same as the final time point.}

\item{NewTrials}{An array indicating the trials for categorical predictions. The array must have dimension \code{M x C x NNewVisits}
and contain only non-negative integers. The default sets \code{NewTrials} to \code{NULL}, which assumes the trials for all predictions
are the same as the final time point.}

\item{type}{A character string indicating the type of prediction, choices include "temporal" and "spatial".}

\item{...}{other arguments.}
}
\value{
\code{predict.spBFA} returns a list containing the following objects.

  \describe{

  \item{\code{MuTauAlpha}}{A \code{list} containing three matrices, \code{mu},
  \code{tau} and \code{alpha}. Each matrix is dimension \code{NKeep x s}, where
  \code{s} is the number of new time points. Each matrix contains posterior
  samples obtained by Bayesian krigging.}

  \item{\code{Y}}{A \code{list} containing \code{s} posterior predictive distribution
  matrices. Each matrix is dimension \code{NKeep x s}, where \code{s}
  is the number of new time points. Each matrix is obtained through Bayesian krigging.}

  }
}
\description{
Predicts future observations from the \code{\link{spBFA}} model.
}
\details{
\code{predict.spBFA} uses Bayesian krigging to predict vectors at future
 time points. The function returns the krigged observed outcomes along with the
 observational level parameters (\code{mu}, \code{tau}, and \code{alpha}).
}
\author{
Samuel I. Berchuck
}
